CRUD / CRUD complete search & edit

+++ src/controllers/videoController.js
@@ -10,15 +10,23 @@ export const home = async (req, res) => {
     }
 } 
 
-export const search = (req, res) => {
-    res.render("search", {pageTitle:"search"});
+export const search = async (req, res) => {
+    const {title} = req.query;
+    let videos = [];
+    if(title){
+        videos = await Video.find({title : {$regex:new RegExp(`^${title}`,"i")}});
+    }
+    return res.render("search", {pageTitle:"search", videos});
 }
 
 //videoRouter
 export const watch = async (req, res) => {
     const {id} = req.params;
+    const videoCheck = await Video.exists({_id:id});
+    if(!videoCheck){
+        return res.render("server-error", {pageTitle:"server-error"});
+    }
     const video = await Video.findById(id);
-    console.log(video);
     res.render("watch", {pageTitle:"watch", video});
 }
 
@@ -36,12 +44,32 @@ export const postUpload = async (req, res) =>{
     res.redirect('/');   
 }
 
-export const edit = (req, res) => {
-    const {id}=req.params;
-    res.render("editVideo", {pageTitle:"editVideo"});
+export const getEdit = async (req, res) => {
+    const {id} = req.params;
+    const videoCheck = await Video.exists({_id:id});
+    if(!videoCheck){
+        return res.render("server-error", {pageTitle:"server-error"});
+    }
+    const video = await Video.findById(id);
+    res.render("editVideo", {pageTitle:"editVideo", video});
 }
+
+export const postEdit = async (req, res) => {
+    const {id}=req.params;
+    const {title, description, hashtags} = req.body;
+    await Video.findByIdAndUpdate(id, {
+        title, description, hashtags : hashtags.split(",").map(word => word.charAt(0) =="#" ? word : `#${word}`),
+    });
+    res.redirect(`/videos/${id}`);
+}
+
+
 export const remove = async (req, res) => {
     const {id} = req.params;
+    const videoCheck = await Video.exists({_id:id});
+    if(!videoCheck){
+        return res.render("server-error", {pageTitle:"server-error"});
+    }
     await Video.deleteOne({ _id : id });  
     res.redirect("/");
 }
